Added separate handler for each file types
